fix: satisfy pkg/interfaces file-count limit after model provider work
Merge ModelProvider types and tests into existing interfaces files so
pkg-file-count stays within the 15-file package limit.
